perf: preallocate partition key strings
Partition keys tend to be approximately the same size each time (and in
the default case, always exactly the same size).

This simple change reduces allocations by pre-sizing the next partition
key string to match that of the previous. This should reduce the number
of allocations needed to grow the string for ~10% throughput increase.