refactor: single entry point for partition cache (#8093)
For #8089 I would like to request each partition only once. Since
internally we store both the sort key and the column ranges in one cache
value anyways, there is no reason to offer two different methods to look
them up.

This only changes the `PartitionCache` interface. The actual lookups are
still separate, but will be changed in a follow-up.