refactor: remove panic on impossible error
Remove the logical complexity of error handling for an error that cannot
occur.

This was an artifact of pre-PR refactoring - the error being returned
SHOULD never be reached, as the only error returned is the "your message
is too big" error, and that's not possible because the message size is
validated in the GossipHandle::broadcast() method before it reaches the
reactor.