fix: Update compactor's use of QueryChunk to return TransitionPartitionId
Also rename PartitionInfo's transition_partition_id to be partition_id
so that it's consistent with the QueryChunk method. We might want to
rename the partition_id field to catalog_partition_id, but for now I
think the types will make compactor usage clear enough.

This gets compactor to compile and pass its tests.