refactor(influxql): make difference a user-defined window function (#8257)
Change the implementation of DIFFERENCE, and NON_NEGATIVE_DIFFERENCE,
to use a user-defined window function, rather than a user-defined
aggregator function. This slightly simplifies the implementation.

The window function calculates difference operation on the full
window in one operation rather than a row at a time. Where ther are
no NULL values in the input this function uses arrow compute kernels.

As part of this refactor the window function's implementation is
moved out of the planner module.